requirements.txt
python-dotenv
langchain
langchain-community
langchain-openai
langchain-chroma
openai
chromadb
tiktoken
pypdf
